WIP: [flang][OpenMP] Lower REDUCTION clause for SECTIONS
The tricky bit here is that we need to generate the reduction symbol
mapping inside each of the nested SECTION constructs. This is a bit
similar to omp.canonical_loop inside of omp.wsloop, except the SECTION
constructs come from the PFT.

To make this work I moved the lowering of the SECTION constructs inside
of the lowering SECTIONS (where reduction information is still
available). This subverts the normal control flow for OpenMP lowering a
bit.

One alternative option I investigated would be to generate the SECTION
CONSTRUCTS as normal as though there were no reduction, and then to fix
them up after control returns back to genSectionsOp. The problem here is
that the code generated for the section body has the wrong symbol
mapping for the reduction variable, so all of the nested code has to
be patched up. In my prototype version this was even more hacky than
what the solution I settled upon.

If you have a better idea please let me know. It would be better if
SECTION(S) didn't get special treatment.

This is draft until I have implemented MLIR to LLVMIR conversion so that
it can be tested properly.