[mlir][OpenMP] don't add compiler-generated barrier in single threaded code
We add barriers to the firstprivate copy region when they are required
to avoid a race condition with the lastprivate clause.

The problem is that these barriers are added by the compiler and I don't
think there is any way for the user to prevent their generation.

An idiom for using taskloop looks like this
```
!$omp parallel
  !$omp single
    !$omp taskloop firstprivate(a) lastprivate(a)
      ...
  !$omp end single
!$omp end parallel
```

This is so that there are multiple threads for the generated tasks to be
distributed over, but we don't generate the tasks afresh in every
thread.

The problem comes when the taskloop requires a barrier to prevent the
datarace between firstprivate and lastprivate. This barrier will then be
generated inside of SINGLE and so only one thread will encounter the
barrier: leading to a deadlock.

This patch works around the problem by detecting this situation
statically and not generating the barrier. There are cases where we
cannot detect this statically (e.g. if the TASKLOOP is inside a function
call inside of SINGLE). The program will still deadlock in this case
after my patch. I'm unsure what the solution would be and want to fix
the common case in LLVM 22 before engaging in a longer discussion as to
whether there is a better way to handle the more general case.

Testing using wsloop because I want to land (or not) independently of
taskloop.